backend/requirements.txt
# Airline Schedule Management System - Backend Requirements

# AI/LLM Framework
langgraph==0.2.45
langchain==0.3.12
langchain-anthropic==0.3.7
langchain-core==0.3.22

# Web Framework
fastapi==0.115.5
uvicorn[standard]==0.32.1
pydantic==2.10.3
python-multipart==0.0.18

# Database
psycopg2-binary==2.9.10
alembic==1.14.0
sqlalchemy==2.0.36

# Neo4j
neo4j==5.27.0

# Utilities
python-dotenv==1.0.1
pyyaml==6.0.2
python-dateutil==2.9.0

# Testing
pytest==8.3.4
pytest-asyncio==0.25.0
pytest-cov==6.0.0
pytest-mock==3.14.0
httpx==0.28.1

# Development
black==24.10.0
pylint==3.3.2
mypy==1.13.0
isort==5.13.2

# Logging & Monitoring
structlog==24.4.0
